restructure tests and fix ktlint
