+++ src/main/kotlin/no/skatteetaten/aurora/gorg/extensions/BuildConfig.kt
@@ -1,12 +1,11 @@
 package no.skatteetaten.aurora.gorg.extensions
 
 import io.fabric8.openshift.api.model.BuildConfig
-import no.skatteetaten.aurora.gorg.service.ApplicationDeploymentResource
 import no.skatteetaten.aurora.gorg.service.BuildConfigResource
 import java.time.Duration
 import java.time.Instant
 
-fun BuildConfig.toResource(now: Instant) : BuildConfigResource {
+fun BuildConfig.toResource(now: Instant): BuildConfigResource {
     val removalTime = this.removalTime()
 
     return BuildConfigResource(
@@ -16,4 +15,3 @@ fun BuildConfig.toResource(now: Instant) : BuildConfigResource {
         removalTime = removalTime
     )
 }
-
