+++ src/main/kotlin/no/skatteetaten/aurora/gorg/extensions/HasMetadata.kt
@@ -7,10 +7,8 @@ import java.time.Instant
 const val REMOVE_AFTER_LABEL = "removeAfter"
 const val TERMINATING_PHASE = "Terminating"
 
-
 fun HasMetadata.removalTime(): Instant {
     return this.metadata.labels[REMOVE_AFTER_LABEL]?.let {
         Instant.ofEpochSecond(it.toLong())
     } ?: throw IllegalStateException("removeAfter is not set or valid timstamp")
 }
-
