+++ src/main/kotlin/no/skatteetaten/aurora/gorg/model/ApplicationDeployment.kt
@@ -38,5 +38,4 @@ data class ApplicationDeployment(
             Instant.ofEpochSecond(it.toLong())
         } ?: throw IllegalStateException("removeAfter is not set or valid timstamp")
     }
-
 }
