+++ src/main/kotlin/no/skatteetaten/aurora/gorg/service/TemporaryDto.kt
@@ -3,10 +3,8 @@ package no.skatteetaten.aurora.gorg.service
 import java.time.Duration
 import java.time.Instant
 
-
-
 interface BaseResource {
-    val ttl:Duration
+    val ttl: Duration
     val name: String
     val removalTime: Instant
 }
@@ -18,7 +16,6 @@ data class ApplicationDeploymentResource(
     override val removalTime: Instant
 ) : BaseResource
 
-
 data class BuildConfigResource(
     override val name: String,
     val namespace: String,
@@ -26,10 +23,8 @@ data class BuildConfigResource(
     override val removalTime: Instant
 ) : BaseResource
 
-
 data class ProjectResource(
     override val name: String,
     override val ttl: Duration,
     override val removalTime: Instant
 ) : BaseResource
-
