+++ src/test/kotlin/no/skatteetaten/aurora/gorg/service/DeleteServiceTest.kt
@@ -1,9 +1,12 @@
 package no.skatteetaten.aurora.gorg.service
 
 import assertk.assert
+import assertk.assertions.isEqualTo
 import assertk.assertions.isFalse
 import assertk.assertions.isTrue
+import no.skatteetaten.aurora.gorg.ApplicationDeploymentBuilder
 import no.skatteetaten.aurora.gorg.ProjectDataBuilder
+import no.skatteetaten.aurora.gorg.extensions.execute
 import no.skatteetaten.aurora.gorg.extensions.toResource
 import org.junit.jupiter.api.Test
 import java.time.Duration
@@ -24,12 +27,26 @@ class DeleteServiceTest : AbstractOpenShiftServerTest() {
     @Test
     fun `Return not deleted for non-existing project`() {
         val deleteService = DeleteService(openShiftServer.openshiftClient, true)
-        val deleted = deleteService.deleteProject(ProjectResource(
-            name = "non-existing-name",
-            ttl = Duration.ZERO,
-            removalTime = Instant.now()))
+        val deleted = deleteService.deleteProject(
+            ProjectResource(
+                name = "non-existing-name",
+                ttl = Duration.ZERO,
+                removalTime = Instant.now()
+            )
+        )
         assert(deleted).isFalse()
     }
 
-
+    @Test
+    fun `delete expired applicationDeployments`() {
+        val deleteService = DeleteService(mockClient, true)
+        val ad = ApplicationDeploymentBuilder().build()
+        val request = mockServer.execute(
+            true
+        ) {
+            assert(deleteService.deleteApplicationDeployment(ad.toResource(Instant.now()))).isTrue()
+        }
+        assert(request.method).isEqualTo("DELETE")
+        assert(request.path).isEqualTo("/apis/skatteetaten.no/v1/namespaces/${ad.metadata.namespace}/applicationdeployments/${ad.metadata.name}")
+    }
 }
